User.id is bigInteger

+++ database/migrations/2017_02_18_145707_create_concepts_table.php
@@ -30,7 +30,7 @@ class CreateConceptsTable extends Migration
 
             $table->string('source_url')->nullable();
 
-            $table->integer('owner_id')->unsigned();
+            $table->bigInteger('owner_id')->unsigned();
             $table->foreign('owner_id')->references('id')->on('users');
 
             $table->timestamps();
